Add all missing binary tools back.
These were all skipped because of the default .gitignore.
But a lot of these are .bat files for building, tools like 7zip for compression, and the binary png and bmp for the UI.  These are unmodified from original source.

This is bad form to have tools here in the same location as the built product, but we'll worry about it later.

Skipping the Final Release\win32 folder, as that looks to be actual build results.  Might need these later.